- add test case

+++ IDD Dialer/WYPopoverController/WYPopoverController.m
@@ -1403,13 +1403,13 @@ static CGFloat edgeSizeFromCornerRadius(CGFloat cornerRadius) {
 {
     UIWindow *keyWindow = [[UIApplication sharedApplication] keyWindow];
     
-    int count = keyWindow.subviews.count;
+    NSUInteger count = keyWindow.subviews.count;
     
     UIView *result = keyWindow;
     
     if (count > 0)
     {
-        int lastIndex = count - 1;
+        NSInteger lastIndex = count - 1;
         
         for (NSInteger i = lastIndex; i >= 0; i--)
         {
